test(e2e): collapse 4 starter-project shard specs into one
Replace starter-projects-shard{1..4}.spec.ts — four near-identical files that each checked one quarter of the starter templates for outdated components — with a single starter-projects-outdated-components.spec.ts that generates the four per-quarter tests from a loop. Coverage and the four-way parallelism are unchanged (still four test blocks the CI shards distribute); the quarter is derived from the test index instead of hard-coded per file. Drops the meaningless shard-N filenames.

Phase 1 (P2). ~378 LOC across four files -> ~103 LOC in one.

Refs LE-1665
